+++ src/a2x_pack_input.p.h
@@ -31,7 +31,7 @@ extern const char* a_input_name(const AInput* Input);
 extern bool a_input_working(const AInput* Input);
 
 extern int a_input_numControllers(void);
-extern void a_input_setController(int Index);
+extern void a_input_setController(size_t Index);
 
 extern bool a_button_get(const AInput* Button);
 extern void a_button_unpress(const AInput* Button);
